core: Place the line rendering algorithm in `renderer.rs` (#22)
More to come to this file in the future.
