+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod canvas;
 pub mod color;
 pub mod geometry;
+mod renderer;
